requirements.txt
Django==3.2.15
gunicorn==20.0.4
psycopg2-binary==2.8.6
celery~=5.2
redis~=4.3
djangorestframework==3.12.4
drf-yasg~=1.21